Start restructuring
- Invalidate most of the readme
- Update the filetree to match the upcoming updates
- Minor updates to the files to preserve the bot functioning
