+++ README.md
@@ -2,6 +2,12 @@
 
 Telegram bot for simple remote control of the device it is running on.
 
+## Redesign in progress
+
+The current readme is mainly outdated.
+
+<s>
+
 ## Requirements / limitations
 
 - To run the bot, on the device you need Python 3 and pip3 and have the libraries
@@ -38,12 +44,12 @@ Supported commands (leading slash can be omitted, lower/upper case do not matter
 
 - `/help` - List available commands (you can find details there)
 
-- `/key [<ARGS>] <KEYS> [<KEYS>...]` (**Xorg only**) - Generate keypress event for a key,
+- `/key [&lt;ARGS&gt;] &lt;KEYS&gt; [&lt;KEYS&gt;...]` (**Xorg only**) - Generate keypress event for a key,
   a shortcut, or a sequence of them. All the arguments are separated by space and forwarded
-  to `xdotool key`, thus, `<KEYS>` must be valid `xdotool` keysequences, and it is possible
-  to specify additional arguments `<ARGS>` (refer to `xdotool key --help` for details)
+  to `xdotool key`, thus, `&lt;KEYS&gt;` must be valid `xdotool` keysequences, and it is possible
+  to specify additional arguments `&lt;ARGS&gt;` (refer to `xdotool key --help` for details)
 
-- `/type <STRING>` (**Xorg only**) - Type the given text on the keyboard through
+- `/type &lt;STRING&gt;` (**Xorg only**) - Type the given text on the keyboard through
   keyboard events.
 
 - `/screen` (**Windows, macOS or Xorg**) - Take a screenshot and send it as a Telegram
@@ -51,16 +57,18 @@ Supported commands (leading slash can be omitted, lower/upper case do not matter
 
 - `/screenf` (-||-) - Just like `/screen`, but sends the screenshot as a document.
 
-- `/run <COMMAND & ARGS SHELL-STYLE>` - run the command without shell but with
+- `/run &lt;COMMAND &amp; ARGS SHELL-STYLE&gt;` - run the command without shell but with
   shell-style arguments splitting (quoting and escaping is supported)
 
-- `/rawrun <COMMAND & ARGS WHITESPACE-SEPARATED>` - run the command without shell
+- `/rawrun &lt;COMMAND &amp; ARGS WHITESPACE-SEPARATED&gt;` - run the command without shell
   and split it by whitespaces, ignoring quotes and backslashes
 
-- `/shell <SHELL COMMAND>` - run the command in shell.
+- `/shell &lt;SHELL COMMAND&gt;` - run the command in shell.
 
 Note: `/exec` and `/rawexec` are synonyms for `/run` and `/rawrun` respectively (for
 backward compatibility).
+
+</s>
 
 ## Security
