+++ Rimokon/util.py
